src/receiver_code/requirements.txt
functions-framework==3.*
google-cloud-storage

